final commit

+++ Project0_PhucHuynh/src/main/java/org/bank/daos/implementation/CustomerDAOImpl.java
@@ -1,5 +1,6 @@
 package org.bank.daos.implementation;
 
+import org.apache.log4j.Logger;
 import org.bank.daos.CustomerDAO;
 import org.bank.exeption.BankException;
 import org.bank.model.Customer;
@@ -12,7 +13,14 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * * This class is created to to act as DOA and communicate with database
+ */
 public class CustomerDAOImpl implements CustomerDAO {
+    private static Logger logger = Logger.getLogger(CustomerDAOImpl.class);
+    /*
+     * This method takes customer object as an input and insert it to the database waiting to get approved abd make sure the username is equine
+     */
     @Override
     public void registerForNewAccount(Customer customer) throws BankException {
         try (Connection connection = ConnectionManager.getConnection()) {
@@ -36,7 +44,7 @@ public class CustomerDAOImpl implements CustomerDAO {
             preparedStatement1.setString(1, customer.getUsername());
             preparedStatement1.setString(2, customer.getPassword());
             preparedStatement1.setString(3, customer.getCustomerName());
-            preparedStatement1.setLong(4, customer.getContact());
+            preparedStatement1.setString(4, customer.getContact());
             preparedStatement1.setString(5, customer.getStatus());
 
             int c = preparedStatement1.executeUpdate();
@@ -47,6 +55,11 @@ public class CustomerDAOImpl implements CustomerDAO {
             System.out.println(e);
         }
     }
+
+    /*
+     * This method takes username and password and compare to the database if there is a match @return true and give the user an access
+     * to the customer UI if not it will ask the user to re-enter
+     */
 
     @Override
     public boolean validateAccount(String username, String password) throws BankException {
@@ -76,29 +89,29 @@ public class CustomerDAOImpl implements CustomerDAO {
         return status;
     }
 
-
-
+    /*
+     * this method that will display play all of the customers that are pending to be approved
+     */
     @Override
     public List<Customer> displayAllPendingCustomer() throws BankException {
         List<Customer> list = new ArrayList<>();
-        try(Connection connection = ConnectionManager.getConnection()) {
+        try (Connection connection = ConnectionManager.getConnection()) {
             String sql = "SELECT customer_id, customer_name, contact, customer_id, status\n" +
                     "FROM mybank_schema.customer where status = ?;";
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
             preparedStatement.setString(1, "Pending");
             ResultSet resultSet = preparedStatement.executeQuery();
 
-            while(resultSet.next()){
+            while (resultSet.next()) {
                 Customer customer = new Customer();
                 customer.setCustomerId(resultSet.getInt("customer_id"));
                 customer.setCustomerName(resultSet.getString("customer_name"));
-                customer.setContact(resultSet.getLong("contact"));
+                customer.setContact(resultSet.getString("contact"));
                 customer.setStatus(resultSet.getString("status"));
                 list.add(customer);
             }
-            if(list.isEmpty()){
-                throw new BankException("There is no pending customer.");
-            }
+            if (list.isEmpty()) throw new BankException("There is no pending customer.");
+
 
         } catch (SQLException throwables) {
             throwables.printStackTrace();
@@ -108,11 +121,14 @@ public class CustomerDAOImpl implements CustomerDAO {
 
         return list;
     }
-
+    /*
+     *  this method takes the customer id and approve the status on the database so they can open the bank accounts with
+     */
     @Override
     public boolean approveCustomerAccountById(int id) {
-        try(Connection connection = ConnectionManager.getConnection()) {
-            String sql ="UPDATE mybank_schema.customer\n" +
+        try (Connection connection = ConnectionManager.getConnection()) {
+            if(id <0) return false;
+            String sql = "UPDATE mybank_schema.customer\n" +
                     "SET status= ? \n" +
                     "WHERE customer_id= ?\n";
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
@@ -126,10 +142,13 @@ public class CustomerDAOImpl implements CustomerDAO {
         }
         return true;
     }
-
+    /*
+     * This method takes customerId and reject the account  by delete it off the database
+     */
     @Override
     public boolean rejectCustomerAccountById(int id) {
-        try(Connection connection = ConnectionManager.getConnection()) {
+        try (Connection connection = ConnectionManager.getConnection()) {
+            if(id < 0)  return false;
             String sql = "DELETE FROM mybank_schema.customer\n" +
                     "WHERE customer_id = ?;";
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
@@ -137,9 +156,9 @@ public class CustomerDAOImpl implements CustomerDAO {
             int c = preparedStatement.executeUpdate();
 
         } catch (SQLException throwables) {
-            throwables.printStackTrace();
+            logger.error(throwables);
         } catch (ClassNotFoundException e) {
-            e.printStackTrace();
+            logger.error(e);
         }
         //should return true when the query is executed successfully
         return true;
